# Collect outputs asynchronously + fix FD leak