Fixes https://github.com/charlesetc/feather/issues/24

Collecting stdout and stderr is now done in separate threads in order to avoid blocking if the program outputs more than the pipe capacity. Events are used to communicate between threads.

Add a test for the issue.