@Firobe 2022-01-11
This PR also fixes another bug linked to a file descriptor leak : this can be observed by repeatedly calling Feather.collect from a single OCaml program.

This will quickly reach the FD limit, while it should not.