@charlesetc 2022-01-11
This change looks good to me - touched things up a bit to match the rest of the style. 

Do you mind running a quick benchmark to make sure this doesn't make things dramatically slower? Aside from that it should be good to go.

Thanks for fixing these bugs!