@Firobe 2022-01-11
Here's a small benchmark, for the following snippet, which should be the worst case: repeatedly launching small collects.

```ocaml
let bench =
  let open Feather in
  for _ = 1 to 1000 do
    process "echo" [ "test" ] |> collect stdout |> ignore
  done
```

Before
```
Benchmark 1: dune exec ./main.exe
  Time (mean ± σ):     513.5 ms ±   1.6 ms    [User: 425.0 ms, System: 96.2 ms]
  Range (min … max):   511.0 ms … 516.5 ms    10 runs
```

After
```
Benchmark 1: dune exec ./main.exe
  Time (mean ± σ):     519.6 ms ±   2.5 ms    [User: 426.9 ms, System: 131.9 ms]
  Range (min … max):   514.3 ms … 522.9 ms    10 runs
```

As expected, the system time increases a bit from creating threads but for the overall runtime we lose approximately 5ms per run, so 5/1000 ms per `collect` call, which is 5 µs. Should be good :)